Remove random_variable abstraction
The cardinality of random variables will be recorded and maintained
by the xt::xarray type.  The client code must specify cardinality
via the data input.  This simplifies the design of pgmfactors with
no real downside.  We may later add an adapter to allow input
of data without the client needed to use xt::xarray explicitly.