Fix factor reduction bug
There was a bug due to a simple error in programming logic which
would cause factor_reduction to throw an error.

The problem was simply that the logic of iterating over an assignment
map stepwise within the for-loop iterating over the input variables was
not compatible.  For example, if two assignment ids occurred numerically
between two input variables ids, then the input variable iterator
would be advanced and skip over a necessary variable.

This commit fixes the bug by changing to logic for a for-loop over a
single iterator to a while-loop over two iterators, advancing each one
explicitly when necessary.

This commit incidently fixes a second problem which could have occurred
if the iterator over the assignments map had been advanced past its end.